feat: componentList -> projectList로 rename, feature디렉토리로 위치 변경, project 관련 타입 선언
