+++ src/features/projectList/index.tsx
@@ -1,0 +1,3 @@
+import ProjectList from "./components/ProjectList";
+
+export default ProjectList
