+++ src/stories/componentList/ExampleCard.stories.ts
@@ -1,4 +1,4 @@
-import { ExampleCard } from "@/components/layout/componentList/card/ExampleCard";
+import { ExampleCard } from "@/features/projectList/components/card/ExampleCard";
 import { bgColors } from "@/resources/colors/colors";
 import type { Meta, StoryObj } from "@storybook/react";
